{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{ cookiecutter.project_name }}"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sonolus.py~=0.1"
]
